Move and split atomic heterogeneous tests (improves build times)
