+++ libcudacxx/test/libcudacxx/heterogeneous/atomic/common.h
@@ -6,13 +6,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-// UNSUPPORTED: nvrtc, pre-sm-60
-// UNSUPPORTED: windows && pre-sm-70
-
 #include <cuda/std/atomic>
 #include <cuda/std/cassert>
 
-#include "helpers.h"
+#include "../helpers.h"
 
 template <int Operand>
 struct store_tester
@@ -186,55 +183,3 @@ public:
 private:
   int array[128];
 };
-
-__host__ __device__ void validate_not_lock_free()
-{
-  cuda::std::atomic<big_not_lockfree_type> test;
-  assert(!test.is_lock_free());
-}
-
-void kernel_invoker()
-{
-  validate_pinned<cuda::std::atomic<signed char>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::std::atomic<signed short>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::std::atomic<signed int>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::std::atomic<signed long>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::std::atomic<signed long long>, arithmetic_atomic_testers>();
-
-  validate_pinned<cuda::std::atomic<unsigned char>, bitwise_atomic_testers>();
-  validate_pinned<cuda::std::atomic<unsigned short>, bitwise_atomic_testers>();
-  validate_pinned<cuda::std::atomic<unsigned int>, bitwise_atomic_testers>();
-  validate_pinned<cuda::std::atomic<unsigned long>, bitwise_atomic_testers>();
-  validate_pinned<cuda::std::atomic<unsigned long long>, bitwise_atomic_testers>();
-
-  validate_pinned<cuda::std::atomic<float>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::std::atomic<double>, arithmetic_atomic_testers>();
-
-  validate_pinned<cuda::std::atomic<big_not_lockfree_type>, basic_testers>();
-
-  validate_pinned<cuda::atomic<signed char, cuda::thread_scope_system>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::atomic<signed short, cuda::thread_scope_system>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::atomic<signed int, cuda::thread_scope_system>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::atomic<signed long, cuda::thread_scope_system>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::atomic<signed long long, cuda::thread_scope_system>, arithmetic_atomic_testers>();
-
-  validate_pinned<cuda::atomic<unsigned char, cuda::thread_scope_system>, bitwise_atomic_testers>();
-  validate_pinned<cuda::atomic<unsigned short, cuda::thread_scope_system>, bitwise_atomic_testers>();
-  validate_pinned<cuda::atomic<unsigned int, cuda::thread_scope_system>, bitwise_atomic_testers>();
-  validate_pinned<cuda::atomic<unsigned long, cuda::thread_scope_system>, bitwise_atomic_testers>();
-  validate_pinned<cuda::atomic<unsigned long long, cuda::thread_scope_system>, bitwise_atomic_testers>();
-
-  validate_pinned<cuda::atomic<float>, arithmetic_atomic_testers>();
-  validate_pinned<cuda::atomic<double>, arithmetic_atomic_testers>();
-
-  validate_pinned<cuda::atomic<big_not_lockfree_type, cuda::thread_scope_system>, basic_testers>();
-}
-
-int main(int arg, char** argv)
-{
-  validate_not_lock_free();
-
-  NV_IF_TARGET(NV_IS_HOST, (kernel_invoker();))
-
-  return 0;
-}
